Separated components into separate files
